# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MaintainIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account (admin or technician)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update the current user's profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "400":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Technicians ----------
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets with search, filter, pagination
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: technicianId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetListResponse"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Asset code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset's full detail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assets/{id}/history:
    get:
      operationId: getAssetHistory
      tags: [assets]
      summary: Get an asset's full history trail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Public ----------
...
  /ai/triage:
    post:
      operationId: triageIssue
      tags: [ai]
      summary: Get AI-assisted issue triage suggestions (no auth, used before final submission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTriageInput"
      responses:
        "200":
          description: AI triage suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTriageResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: AI provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Issues ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---------- Users / Auth ----------
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, technician] }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [admin, technician] }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        avatarUrl: { type: string }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 6 }

    # ---------- Technicians ----------
...
    Asset:
      type: object
      required:
        [
          id,
          name,
          assetCode,
          category,
          location,
          condition,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        assetCode: { type: string }
        category: { type: string }
        location: { type: string }
        model: { type: ["string", "null"] }
        manufacturer: { type: ["string", "null"] }
        purchaseDate: { type: ["string", "null"], format: date }
        warrantyExpiry: { type: ["string", "null"], format: date }
        condition: { type: string, enum: [excellent, good, fair, poor] }
        status:
          type: string
          enum: [operational, under_maintenance, out_of_service, retired]
        assignedTechnicianId: { type: ["integer", "null"] }
        assignedTechnicianName: { type: ["string", "null"] }
        lastServiceDate: { type: ["string", "null"], format: date }
        nextServiceDate: { type: ["string", "null"], format: date }
        internalNotes: { type: ["string", "null"] }
        publicUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssetInput:
      type: object
      required: [name, assetCode, category, location]
      properties:
        name: { type: string, minLength: 1 }
        assetCode: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        model: { type: string }
        manufacturer: { type: string }
        purchaseDate: { type: string, format: date }
        warrantyExpiry: { type: string, format: date }
        condition: { type: string, enum: [excellent, good, fair, poor] }
        status:
          type: string
          enum: [operational, under_maintenance, out_of_service, retired]
        assignedTechnicianId: { type: integer }
        lastServiceDate: { type: string, format: date }
        nextServiceDate: { type: string, format: date }
        internalNotes: { type: string }
    AssetUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        location: { type: string }
        model: { type: string }
        manufacturer: { type: string }
        purchaseDate: { type: string, format: date }
        warrantyExpiry: { type: string, format: date }
        condition: { type: string, enum: [excellent, good, fair, poor] }
        status:
          type: string
          enum: [operational, under_maintenance, out_of_service, retired]
        assignedTechnicianId: { type: ["integer", "null"] }
        lastServiceDate: { type: string, format: date }
        nextServiceDate: { type: string, format: date }
        internalNotes: { type: string }
    AssetListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    AiTriageInput:
      type: object
      required: [assetCode, title, description, category, priority]
      properties:
        assetCode: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        category: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
    AiTriageResult:
      type: object
      required:
        [
          suggestedTitle,
          suggestedCategory,
          suggestedPriority,
          possibleCauses,
          diagnosticChecks,
          safetyNotes,
        ]
      properties:
        suggestedTitle: { type: string }
        suggestedCategory: { type: string }
        suggestedPriority: { type: string, enum: [low, medium, high, critical] }
        possibleCauses:
          type: array
          items: { type: string }
        diagnosticChecks:
          type: array
          items: { type: string }
        safetyNotes: { type: string }
        recurringWarning: { type: ["string", "null"] }
...
    HistoryEntry:
      type: object
      required: [id, assetId, assetName, userName, action, createdAt]
      properties:
        id: { type: integer }
        assetId: { type: integer }
        assetName: { type: string }
        issueId: { type: ["integer", "null"] }
        issueNumber: { type: ["string", "null"] }
        userId: { type: ["integer", "null"] }
        userName: { type: string }
        action: { type: string }
        status: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    HistoryListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }

    # ---------- Dashboard ----------
...