# lib/api-zod/src/generated/types/aiTriageResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MaintainIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiTriageResultSuggestedPriority } from './aiTriageResultSuggestedPriority';

export interface AiTriageResult {
  suggestedTitle: string;
  suggestedCategory: string;
  suggestedPriority: AiTriageResultSuggestedPriority;
  possibleCauses: string[];
  diagnosticChecks: string[];
  safetyNotes: string;
  /** @nullable */
  recurringWarning?: string | null;
}